Added ExperiencesFragmentTests
